requirements.txt
# Oronym & Lyric Assistant - Dependencies
# Python 3.11+ supported

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.3
pydantic-settings>=2.2.1

# Database
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
alembic>=1.13.1
greenlet>=3.0.3

# Cache
redis>=5.0.3

# Machine Learning
torch>=2.2.0
transformers>=4.38.0
faiss-cpu>=1.8.0
gensim>=4.3.2

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.12.0
numpy>=1.26.0

# G2P Fallback
g2p-en>=2.1.0

# HTTP Client
httpx>=0.27.0
aiofiles>=23.2.1

# Monitoring
prometheus-client>=0.20.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# Code Quality
black>=24.3.0
ruff>=0.3.0

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.1

sentencepiece>=0.1.99
faiss-cpu>=1.7.4
